initial commit for moving six mpi_runners regression tests into pytest framework. renames test_make_runners to be more accurate. adds initial test_misc_features
