+++ libensemble/tests/unit_tests/test_misc_features.py
@@ -1,0 +1,24 @@
+import time
+import pytest
+import platform
+import numpy as np
+
+import libensemble.tests.unit_tests.setup as setup
+
+
+@pytest.mark.extra
+def test_profiling():
+    pass
+
+@pytest.mark.extra
+def test_deliberate_error():
+    pass
+
+@pytest.mark.extra
+def test_worker_exception():
+    pass
+
+if __name__ == "__main__":
+    test_term_test_1()
+    test_term_test_2()
+    test_term_test_3()
